+++ src/interlegis/portalmodelo/policy/tests/test_frontpage.py
@@ -57,10 +57,12 @@ class FrontPageTestCase(unittest.TestCase):
         self.assertIn('navTreeItem visualNoMarker section-pesquisar-no-lexml', browser.contents)
 
         # Transparência
+#        self.assertIn('navTreeItem visualNoMarker navTreeFolderish section-indice-de-transparencia', browser.contents)
         self.assertIn('navTreeItem visualNoMarker navTreeFolderish section-orcamento-e-financas', browser.contents)
         self.assertIn('navTreeItem visualNoMarker navTreeFolderish section-licitacoes-e-contratos', browser.contents)
         self.assertIn('navTreeItem visualNoMarker navTreeFolderish section-recursos-humanos', browser.contents)
         self.assertIn('navTreeItem visualNoMarker navTreeFolderish section-parlamentares-e-gabinetes', browser.contents)
+        self.assertIn('navTreeItem visualNoMarker navTreeFolderish section-atos-administrativos', browser.contents)
         self.assertIn('navTreeItem visualNoMarker section-acesso-a-informacao', browser.contents)
         self.assertIn('navTreeItem visualNoMarker section-dados-abertos', browser.contents)